+++ frontend/src/components/PriorityMatrix/card/TaskCard.tsx
@@ -3,8 +3,8 @@
 import { useDraggable } from '@dnd-kit/core';
 import { Calendar } from 'lucide-react';
 import type { Task } from '@/types/task';
-import { CategoryBadge } from '@/components/PriorityMatrix/common/CategoryBadge';
-import { TypeBadge } from '@/components/PriorityMatrix/common/TypeBadge';
+import { CategoryBadge } from '@/components/PriorityMatrix/filter/CategoryBadge';
+import { TypeBadge } from '@/components/PriorityMatrix/filter/TypeBadge';
 
 interface TaskCardProps {
   task: Task;
