+++ frontend/src/components/PriorityMatrix/filter/CategoryManager.tsx
